fix(tcpip): bound IP fragment reassembly and L4/encap parsing (#201)
Harden the parsing and reassembly paths listed in issue #201 against
malformed or truncated captures:

- IPv4 fragment reassembly: validate ihl/tot_len/fragment extent against
  the captured length before deriving the payload or subtracting lengths
  (F-BUG-017, F-BUG-037); check every datagram/hole/realloc allocation
  and keep the old buffer on realloc failure.
- IPv6 fragment reassembly: same hardening plus 32-bit extension-header
  offset accumulators and fragment-hole offsets so long chains cannot
  wrap past 65535 (F-BUG-040); guard payload_len - ext_header_len
  against underflow; validate the post-classification captured length
  (F-BUG-036).
- ip_streams fragment map: cap at MMT_IP_FRAG_MAP_MAX_ENTRIES (1024)
  with stalest-entry eviction, sweep entries idle >=
  MMT_IP_FRAG_TIMEOUT_SEC (30s) from the existing session-expiry timer
  pass, and drain all remaining datagrams on handler close via new
  mmt_handler frag_map_sweep_fct/frag_map_drain_fct hooks armed by the
  TCP/IP plugin on the first fragment (F-BUG-020). Malformed fragments
  are never parked in the map.
- hashmap: track nkeys on insert/remove/cleanup and let hashmap_walk
  callbacks remove the current entry safely (successor cached first).
- TCP segments: clamp tcp_seg_reassembly copies to the destination
  budget (F-BUG-019) and tag arena-backed nodes so tcp_seg_free /
  tcp_seg_free_list never call free() on arena memory (F-BUG-039).
- mmt_bytestream_to_number: check the byte budget before dereferencing
  (F-BUG-024); ICMP data extraction copies at most the captured length
  (F-BUG-018); SCTP chunk walking validates offsets, minimum chunk
  length and 4-byte alignment (F-BUG-034); GRE base and optional fields
  are checked against caplen (F-BUG-035); GTP optional/next-extension
  header reads are bounds-checked (F-BUG-021); VLAN/QinQ tag reads are
  bounds-checked (F-BUG-106); ip4/ip6 id objects are removed from
  ips_map before being freed on error paths (F-BUG-025).

Adds tools/phase0/tests/ip_frag_bounds_test.c (+ runner) exercising all
of the above under ASan/UBSan with crafted packets: 19 checks.

+++ src/mmt_tcpip/lib/protocols/proto_ip_frag.h
@@ -10,8 +10,11 @@
 
 struct ip_frag {
    LIST_ENTRY(ip_frag) frags;  // sibling fragments
-   uint16_t loff;              // leftmost offset in reassembly buffer
-   uint16_t roff;              // rightmost offset in reassembly buffer
+   /* Issue #201 (F-BUG-040): hole boundaries were uint16_t — an attacker-
+    * controlled off+len could wrap past 65535 and corrupt hole tracking.
+    * 32-bit fields make the (bounded) arithmetic exact. */
+   uint32_t loff;              // leftmost offset in reassembly buffer
+   uint32_t roff;              // rightmost offset in reassembly buffer
 };
